Add mime_type to attachment
This isn't necessarily required by the API. But when explicitly set to
"application/vnd.google-apps.document", you have an option on the event
to use that attachment as meeting notes.

Ideally there would be no manual step at all to set an attachment as
meeting notes. There doesn't seem to be a way to do this though.